enhancement and start to add dashboard admin
