--- public/js/navItemActive.js
@@ -1,3 +1,0 @@
-let page = location.pathname.split('/').slice(-1);
-$('nav ul li').removeClass('active');
-$('#'+ page).addClass('active');
