+++ src/Controller/HomeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\User;
 use Core\DefaultAbstract\DefaultAbstractController;
 use Core\Exception\CoreException;
 
@@ -19,8 +20,10 @@ class HomeController extends DefaultAbstractController
      */
     public function indexAction()
     {
-        if (isset($_SESSION['logged']) === true) {
-            $this->redirectTo('Admin/userAdmin');
+        if ($this->isLogged()) {
+            $user = $this->getUserLogged();
+            assert($user instanceof User);
+            $this->redirectTo('user');
         }
 
         $this->renderView(
