+++ template/back/admin/Comment/comment.html.twig
@@ -32,7 +32,7 @@
         </div>
     </div>
     <div class="p-3 mt-4 mb-5 mx-auto w-50 border-top border-bottom border-info">
-        <a href="/article" class="m-2 p-2 btn btn-primary">
+        <a href="../../../../index.php" class="m-2 p-2 btn btn-primary">
             <span class="fas fa-angle-double-left"></span>
             Retour à la liste des billets
         </a>
