+++ template/back/admin/article/formArticle.html.twig
@@ -1,4 +1,4 @@
-{% extends "back/layoutBack.html.twig" %}
+{% extends "back/admin/layoutAdmin.html.twig" %}
 {% block content %}
     <h3 class="mb-5">
         <span class="fas fa-feather-alt"></span>
@@ -40,7 +40,7 @@
         </form>
     </div>
     <div class="p-3 mt-4 mb-5 mx-auto w-50 border-top border-bottom border-info">
-        <a href="/article" class="btn btn-primary p-2 my-2">
+        <a href="../../../../index.php" class="btn btn-primary p-2 my-2">
             <span class="fas fa-angle-double-left"></span>
             Retour à la liste des billets</a>
     </div>
