+++ template/back/admin/layoutAdmin.html.twig
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="fr">
+<head>
+    {{ include('partial/_head.html.twig') }}
+</head>
+<body class="p-0 bg-light">
+<header>
+    {{ include('back/admin/partial/_nav.html.twig') }}
+</header>
+<main class="m-4 p-2 text-center">
+    {% block content %}{% endblock %}
+</main>
+<footer class="fixed-bottom bg-dark text-secondary">
+    {{ include('back/admin/partial/_footerAdmin.html.twig') }}
+</footer>
+</body>
+</html>
