+++ template/back/admin/partial/_footerAdmin.html.twig
@@ -1,5 +1,5 @@
 <div class="row d-flex pt-3 px-2">
-    <div class="col-md-6 mb-1">
+    <div class="col mb-1">
         <!--Copyright-->
         <p class="text-center h6">
             <span class="fa fa-copyright"></span>
@@ -7,15 +7,6 @@
             <br>
             Dans le cadre de ma formation avec OpenClassrooms.
         </p>
-    </div>
-    <div class="col-md-6">
-        <!-- Connexion -->
-        <div class="text-center pt2 pb-3">
-            <a href="/Admin/userAdmin" class="btn btn-outline-success">
-                <span class="fas fa-power-off"></span>
-                Administration
-            </a>
-        </div>
     </div>
 </div>
 <!-- script jquery -->
