+++ template/back/admin/partial/_nav.html.twig
@@ -9,28 +9,23 @@
     <div class="collapse navbar-collapse justify-content-end text-uppercase" id="navbarSupportedContent">
         <ul class="navbar-nav">
             <li id="userAdmin" class="nav-item">
-                <a class="nav-link" href="/home">
-                    <i class="fas fa-home" style="font-size:20px"></i>
+                <a class="nav-link" href="/Admin/userAdmin">
+                    <span class="fas fa-tachometer-alt"></span>
                     <span class="sr-only">(current)</span>
-                    Accueil
+                    Dashboard
                 </a>
             </li>
             <li id="article" class="nav-item">
-                <a class="nav-link" href="/article">
+                <a class="nav-link" href="/article/add">
                     <span class="far fa-newspaper"></span>
-                    Blog
+                    Ajouter un article
                 </a>
             </li>
-            <li id="contact" class="nav-item">
-                <a class="nav-link" href="/contact">
-                    <span class="fas fa-envelope-open-text"></span>
-                    Contact
-                </a>
-            </li>
-            <li id="profile" class="nav-item">
-                <a class="nav-link" href="/user/profile">
-                    <span class="fas fa-user"></span>
-                    Profil
+            <li id="userAdmin" class="nav-item">
+                <a class="nav-link" href="/home">
+                    <span class="fas fa-home" style="font-size:20px"></span>
+                    <span class="sr-only">(current)</span>
+                    Accueil
                 </a>
             </li>
             <li><a class="nav-link" href="/authentication/logout">
