+++ template/back/article/articleById.html.twig
@@ -1,99 +1,18 @@
 {% extends "back/layoutBack.html.twig" %}
 {% block content %}
-    <div class="container mx-auto">
-        <div class="p-3 mt-5 rounded bg-blue shadow">
-            <h2>{{ article.title }}</h2>
-            <div class="p-2 my-3 mx-auto w-50 border-top border-bottom border-info">
-                <p class="h6 p-2 text-secondary font-italic">
-                    <span class="h5 font-weight-bold text-secondary font-italic">Auteur :</span>
-                    {{ article.author }}
-                    <br class="mb-2">
-                    <span class="font-weight-bold">Créer le : </span>{{ article.createdAt }}
-                    <br>
-                    <span class="font-weight-bold">Dernière mise à jour : </span>{{ article.updatedAt }}
-                </p>
-            </div>
-            <div class="p-2 mt-3 mx-2 bg-light rounded-top">
-                <p class="h4 font-weight-bold">
-                    {{ article.hat }}
-                </p>
-            </div>
-            <div class="p-3 mb-3 mx-2 bg-white text-dark rounded-bottom">
-                <p class="h4">
-                    {{ article.content }}
-                </p>
-            </div>
-            <div class="row d-flex p-2 my-2 mx-auto">
-                <div class="col-md-6 text-center py-2">
-                    <a href="/article/edit?articleId={{ article.id }}" class="p-2 btn btn-info">
-                        <span class="fas fa-edit"></span>
-                        Éditer
-                    </a>
-                </div>
-                <div class="col-md-6 text-center py-2">
-                    <a href="/article/delete?articleId={{ article.id }}" class="p-2 btn btn-danger">
-                        <span class="fas fa-trash-alt"></span>
-                        Supprimer
-                    </a>
-                </div>
-            </div>
+    {{ include('partial/_articleById.html.twig') }}
+    <div class="row d-flex p-3 mt-4 mb-5 mx-auto w-75 border-top border-bottom border-info">
+        <div class="col-md-6 text-center py-2">
+            <a href="/comment/add?articleId={{ article.id }}" class="p-2 btn btn-info">
+                <span class="fas fa-feather-alt"></span>
+                Écrire un nouveau commetaire
+            </a>
         </div>
-        {% if comments %}
-            <div class="panel-group bg-light shadow rounded" id="accordion" role="tablist" aria-multiselectable="true">
-                <div class="panel panel-default">
-                    <div class="panel-heading p-2 mt-3 mx-auto"
-                         role="tab" id="headingOne">
-                        <h4 class="panel-title">
-                            <a class="text-decoration-none text-info font-weight-bold" role="button"
-                               data-toggle="collapse"
-                               data-parent="#accordion" href="#collapseOne"
-                               aria-expanded="true" aria-controls="collapseOne">
-                                <span class="fas fa-angle-double-down"></span>
-                                Voir les commentaires
-                                <span class="fas fa-angle-double-down"></span>
-                            </a>
-                        </h4>
-                    </div>
-                    <div id="collapseOne" class="panel-collapse collapse in" role="tabpanel"
-                         aria-labelledby="headingOne">
-                        <div class="panel-body">
-                            {% for comment in comments %}
-                                <div class="p-3 my-3 rounded bg-blue shadow">
-                                    <p class="h6 p-2 text-secondary font-italic">
-                                        <span class="h5 font-weight-bold text-secondary font-italic">Auteur :</span>
-                                        {{ comment.author }}
-                                        <br class="mb-2">
-                                        <span class="font-weight-bold">Créer le : </span>{{ comment.createdAt }}<br>
-                                        <span class="font-weight-bold">Dernière mise à jour : </span>{{ comment.updatedAt }}
-                                    </p>
-                                    <p class="h5 p-3 mb-3 mx-2 bg-white rounded">
-                                        {{ comment.content }}
-                                    </p>
-                                    <a href="/comment/see?commentId={{ comment.id }}" class="p-2 my-3 btn btn-primary">
-                                        Voir le commentaire
-                                        <span class="fas fa-angle-double-right"></span>
-                                    </a>
-                                    <br>
-                                </div>
-                            {% endfor %}
-                        </div>
-                    </div>
-                </div>
-            </div>
-        {% endif %}
-        <div class="row d-flex p-3 mt-4 mb-5 mx-auto w-75 border-top border-bottom border-info">
-            <div class="col-md-6 text-center py-2">
-                <a href="/comment/add?articleId={{ article.id }}" class="p-2 btn btn-info">
-                    <span class="fas fa-feather-alt"></span>
-                    Écrire un nouveau commetaire
-                </a>
-            </div>
-            <div class="col-md-6 text-center py-2">
-                <a href="/article" class="p-2 btn btn-primary">
-                    <span class="fas fa-angle-double-left"></span>
-                    Retour à la liste des billets
-                </a>
-            </div>
+        <div class="col-md-6 text-center py-2">
+            <a href="/article" class="p-2 btn btn-primary">
+                <span class="fas fa-angle-double-left"></span>
+                Retour à la liste des billets
+            </a>
         </div>
     </div>
 {% endblock %}
