+++ template/back/comment/formComment.html.twig
@@ -30,6 +30,7 @@
     <div class="p-3 mt-4 mb-5 mx-auto w-50 border-top border-bottom border-info">
         <a href="/article" class="btn btn-primary p-2 my-2">
             <span class="fas fa-angle-double-left"></span>
-            Retour à la liste des billets</a>
+            Retour à la liste des billets
+        </a>
     </div>
 {% endblock %}
