+++ template/back/profile/profile.html.twig
@@ -20,20 +20,6 @@
                     <span class="font-weight-bold">mon adresse mail : </span>{{ user.mail }}
                 </div>
             </div>
-            <div class="row d-flex p-2 my-2 mx-auto">
-                <div class="col-md-6 text-center py-2">
-                    <a href="/Admin/userAdmin/edit?userId={{ user.id }}" class="p-2 btn btn-info">
-                        <span class="fas fa-edit"></span>
-                        Éditer
-                    </a>
-                </div>
-                <div class="col-md-6 text-center py-2">
-                    <a href="/Admin/userAdmin/delete?userId={{ user.id }}" class="p-2 btn btn-danger">
-                        <span class="fas fa-trash-alt"></span>
-                        Supprimer
-                    </a>
-                </div>
-            </div>
         </div>
     </div>
     <div class="p-3 mt-4 mb-5 mx-auto w-50 border-top border-bottom border-info">
