+++ template/front/article/articleById.html.twig
@@ -1,0 +1,10 @@
+{% extends "layout.html.twig" %}
+{% block content %}
+    {{ include('partial/_articleById.html.twig') }}
+    <div class="p-3 mt-4 mb-5 mx-auto w-50 border-top border-bottom border-info">
+        <a href="/blog" class="p-2 btn btn-primary">
+            <span class="fas fa-angle-double-left"></span>
+            Retour à la liste des billets
+        </a>
+    </div>
+{% endblock %}
