+++ template/front/article/articles.html.twig
@@ -1,4 +1,4 @@
-{% extends "back/layoutBack.html.twig" %}
+{% extends "layout.html.twig" %}
 {% block content %}
     {{ include('partial/_article.html.twig') }}
 {% endblock %}
