+++ template/partial/_nav.html.twig
@@ -15,9 +15,30 @@
                     Accueil
                 </a>
             </li>
-            <li id="contact" class="nav-item"><a class="nav-link" href="/contact">Contact</a></li>
-            <li id="register" class="nav-item"><a class="nav-link" href="/register">Inscription</a></li>
-            <li id="authentication" class="nav-item"><a class="nav-link" href="/authentication">Connexion</a></li>
+            <li id="blog" class="nav-item">
+                <a class="nav-link" href="/blog">
+                    <span class="far fa-newspaper"></span>
+                    Blog
+                </a>
+            </li>
+            <li id="contact" class="nav-item">
+                <a class="nav-link" href="/contact">
+                    <span class="fas fa-envelope-open-text"></span>
+                    Contact
+                </a>
+            </li>
+            <li id="register" class="nav-item">
+                <a class="nav-link" href="/register">
+                    <span class="fas fa-file-signature"></span>
+                    Inscription
+                </a>
+            </li>
+            <li id="authentication" class="nav-item">
+                <a class="nav-link" href="/authentication">
+                    <span class="fas fa-sign-in-alt"></span>
+                    Connexion
+                </a>
+            </li>
         </ul>
     </div>
 </nav>
